refactor: lifting state cards up
